+++ src/simsopt/field/tracing.py
@@ -4,7 +4,7 @@ import simsoptpp as sopp
 import logging
 from simsopt.field.magneticfield import MagneticField
 from simsopt.field.sampling import draw_uniform_on_curve, draw_uniform_on_surface
-from simsopt.geo.surface import signed_distance_from_surface
+from simsopt.geo.surface import SurfaceClassifier
 from simsopt.util.constants import ALPHA_PARTICLE_MASS, ALPHA_PARTICLE_CHARGE, FUSION_ALPHA_PARTICLE_ENERGY
 from nptyping import NDArray, Float
 
@@ -354,80 +354,6 @@ def particles_to_vtk(res_tys, filename):
     data = np.concatenate([i*np.ones((res_tys[i].shape[0], )) for i in range(len(res_tys))])
     polyLinesToVTK(filename, x, y, z, pointsPerLine=ppl, pointData={'idx': data})
 
-
-class SurfaceClassifier():
-    r"""
-    Takes in a toroidal surface and constructs an interpolant of the signed distance function
-    :math:`f:R^3\to R` that is positive inside the volume contained by the surface,
-    (approximately) zero on the surface, and negative outisde the volume contained by the surface.
-    """
-
-    def __init__(self, surface, p=1, h=0.05):
-        """
-        Args:
-            surface: the surface to contruct the distance from.
-            p: degree of the interpolant
-            h: grid resolution of the interpolant
-        """
-        gammas = surface.gamma()
-        r = np.linalg.norm(gammas[:, :, :2], axis=2)
-        z = gammas[:, :, 2]
-        rmin = max(np.min(r) - 0.1, 0.)
-        rmax = np.max(r) + 0.1
-        zmin = np.min(z) - 0.1
-        zmax = np.max(z) + 0.1
-
-        self.zrange = (zmin, zmax)
-        self.rrange = (rmin, rmax)
-
-        nr = int((self.rrange[1]-self.rrange[0])/h)
-        nphi = int(2*np.pi/h)
-        nz = int((self.zrange[1]-self.zrange[0])/h)
-
-        def fbatch(rs, phis, zs):
-            xyz = np.zeros((len(rs), 3))
-            xyz[:, 0] = rs * np.cos(phis)
-            xyz[:, 1] = rs * np.sin(phis)
-            xyz[:, 2] = zs
-            return list(signed_distance_from_surface(xyz, surface))
-
-        rule = sopp.UniformInterpolationRule(p)
-        self.dist = sopp.RegularGridInterpolant3D(
-            rule, [rmin, rmax, nr], [0., 2*np.pi, nphi], [zmin, zmax, nz], 1, True)
-        self.dist.interpolate_batch(fbatch)
-
-    def evaluate(self, xyz):
-        rphiz = np.zeros_like(xyz)
-        rphiz[:, 0] = np.linalg.norm(xyz[:, :2], axis=1)
-        rphiz[:, 1] = np.mod(np.arctan2(xyz[:, 1], xyz[:, 0]), 2*np.pi)
-        rphiz[:, 2] = xyz[:, 2]
-        d = np.zeros((xyz.shape[0], 1))
-        self.dist.evaluate_batch(rphiz, d)
-        return d
-
-    def to_vtk(self, filename, h=0.01):
-        from pyevtk.hl import gridToVTK
-
-        nr = int((self.rrange[1]-self.rrange[0])/h)
-        nphi = int(2*np.pi/h)
-        nz = int((self.zrange[1]-self.zrange[0])/h)
-        rs = np.linspace(self.rrange[0], self.rrange[1], nr)
-        phis = np.linspace(0, 2*np.pi, nphi)
-        zs = np.linspace(self.zrange[0], self.zrange[1], nz)
-
-        R, Phi, Z = np.meshgrid(rs, phis, zs)
-        X = R * np.cos(Phi)
-        Y = R * np.sin(Phi)
-        Z = Z
-
-        RPhiZ = np.zeros((R.size, 3))
-        RPhiZ[:, 0] = R.flatten()
-        RPhiZ[:, 1] = Phi.flatten()
-        RPhiZ[:, 2] = Z.flatten()
-        vals = np.zeros((R.size, 1))
-        self.dist.evaluate_batch(RPhiZ, vals)
-        vals = vals.reshape(R.shape)
-        gridToVTK(filename, X, Y, Z, pointData={"levelset": vals})
 
 
 class LevelsetStoppingCriterion(sopp.LevelsetStoppingCriterion):
